dataflow/requirements.txt
apache-beam[gcp]>=2.54.0
google-cloud-firestore>=2.14.0
psycopg2-binary>=2.9.9
geopy>=2.4.1
python-dotenv>=1.0.1
google-cloud-secret-manager>=2.17.0